// common/generated/motion.ts
// Generated by Style Dictionary. Do not edit directly.

export const motion = {
  "duration": {
    "100": "120ms",
    "200": "200ms",
    "400": "400ms",
    "800": "800ms",
    "normal": "200ms",
    "quick": "120ms",
    "slow": "320ms"
  },
  "easing": {
    "ease-in-out": "cubic-bezier(0.4, 0, 0.2, 1)",
    "ease-out": "cubic-bezier(0.16, 1, 0.3, 1)",
    "emph": "cubic-bezier(0.05, 0.7, 0.1, 1)",
    "linear": "cubic-bezier(0.0, 0.0, 1.0, 1.0)",
    "standard": "cubic-bezier(0.2, 0, 0, 1)"
  },
  "spring": {
    "medium": "stiffness(170) damping(22) mass(1)",
    "soft": "stiffness(120) damping(18) mass(1)",
    "strong": "stiffness(220) damping(28) mass(1)"
  }
} as const;

export type MotionDurationToken = keyof typeof motion.duration;

export type MotionEasingToken = keyof typeof motion.easing;

export type MotionSpringToken = keyof typeof motion.spring;